+++ modules/Discipline/CategoryBreakdownTime.php
@@ -103,6 +103,20 @@ if ( ! empty( $_REQUEST['category_id'] ) )
 		}
 	}
 
+	$months = array(
+		'1' => mb_substr( _( 'January' ), 0, 3 ),
+		'2' => mb_substr( _( 'February' ), 0, 3 ),
+		'3' => mb_substr( _( 'March' ), 0, 3 ),
+		'4' => mb_substr( _( 'April' ), 0, 3 ),
+		'5' => mb_substr( _( 'May' ), 0, 3 ),
+		'6' => mb_substr( _( 'June' ), 0, 3 ),
+		'7' => mb_substr( _( 'July' ), 0, 3 ),
+		'8' => mb_substr( _( 'August' ), 0, 3 ),
+		'9' => mb_substr( _( 'September' ), 0, 3 ),
+		'10' => mb_substr( _( 'October' ), 0, 3 ),
+		'11' => mb_substr( _( 'November' ), 0, 3 ),
+		'12' => mb_substr( _( 'December' ), 0, 3 ),
+	);
 
 	$extra = array();
 
@@ -142,7 +156,7 @@ if ( ! empty( $_REQUEST['category_id'] ) )
 				$tf = str_pad( ( $i%12 == 0 ? 12 : $i%12 ), 2, '0', STR_PAD_LEFT );
 
 				//FJ add translation
-				$chart['chart_data'][ $index ][0] = _( ucwords( mb_strtolower( MonthNWSwitch( str_pad( $i%12, 2, '0', STR_PAD_LEFT ), 'tochar' ) ) ) );
+				$chart['chart_data'][ $index ][0] = $months[ (int) $tf ];
 			}
 			else // SYEAR
 			{
@@ -186,7 +200,7 @@ if ( ! empty( $_REQUEST['category_id'] ) )
 				$tf = str_pad( ( $i%12 == 0 ? 12 : $i%12 ), 2, '0', STR_PAD_LEFT );
 
 				//FJ add translation
-				$chart['chart_data'][ $index ][0] = _( ucwords( mb_strtolower( MonthNWSwitch( str_pad( $i%12, 2, '0', STR_PAD_LEFT ), 'tochar' ) ) ) );
+				$chart['chart_data'][ $index ][0] = $months[ (int) $tf ];
 			}
 			else // SYEAR
 			{
@@ -241,7 +255,7 @@ if ( ! empty( $_REQUEST['category_id'] ) )
 				$tf = str_pad( ( $i%12 == 0 ? 12 : $i%12 ), 2, '0', STR_PAD_LEFT );
 
 				//FJ add translation
-				$chart['chart_data'][ $index ][0] = _( ucwords( mb_strtolower( MonthNWSwitch( str_pad( $i%12, 2, '0', STR_PAD_LEFT ), 'tochar' ) ) ) );
+				$chart['chart_data'][ $index ][0] = $months[ (int) $tf ];
 			}
 			else // SYEAR
 			{
@@ -282,7 +296,7 @@ if ( ! empty( $_REQUEST['category_id'] ) )
 				$tf = str_pad( ( $i%12 == 0 ? 12 : $i%12 ), 2, '0', STR_PAD_LEFT );
 
 				//FJ add translation
-				$chart['chart_data'][ $index ][0] = _( ucwords( mb_strtolower( MonthNWSwitch( str_pad( $i%12, 2, '0', STR_PAD_LEFT ), 'tochar' ) ) ) );
+				$chart['chart_data'][ $index ][0] = $months[ (int) $tf ];
 			}
 			else // SYEAR
 			{